Refactor game loop bootstrap helpers
