+++ src/heart/runtime/launcher.py
@@ -1,13 +1,13 @@
-from lagom import Container
+"""Runtime bootstrap helpers for building the game loop."""
 
 from heart.device.selection import select_device
-from heart.runtime.container import build_runtime_container
+from heart.runtime.container import RuntimeContainer, build_runtime_container
 from heart.runtime.game_loop import GameLoop
 from heart.runtime.render_pipeline import RendererVariant
 from heart.utilities.env import Configuration
 
 
-def build_game_loop_container(*, x11_forward: bool) -> Container:
+def build_game_loop_container(*, x11_forward: bool) -> RuntimeContainer:
     render_variant = RendererVariant.parse(Configuration.render_variant())
     device = select_device(x11_forward=x11_forward)
     return build_runtime_container(
